+++ EMP.DataAccess/Repos/Sts/AspnetDeptManagerRepository.cs
@@ -1,15 +1,14 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
 using EMP.Common.Tasks;
 using EMP.Data.Models.Sts;
 using EMP.Data.Repos;
 using EMP.DataAccess.Context;
-using EMP.DataAccess.EFCore;
 using EMP.DataAccess.Repos.Extension;
-using Microsoft.EntityFrameworkCore;
 
-namespace EMP.DataAccess.Repos
+namespace EMP.DataAccess.Repos.Sts
 {
     public class AspnetDeptManagerRepository : IRepository<AspnetDeptManager>
     {
